@@ -3,15 +3,17 @@ using GraphQL.Types;
 using GraphQlWebCore.DependencyInjection;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace GraphQlWebCore
 {
     public class Startup
     {
-        private readonly IHostingEnvironment _env;
+        private readonly IWebHostEnvironment _env;
 
-        public Startup(IHostingEnvironment env)
+        public Startup(IWebHostEnvironment env)
         {
             _env = env;
         }
@@ -25,8 +27,7 @@ namespace GraphQlWebCore
                 .AddMvc();
 
             services
-                .AddMvcCore()
-                .AddJsonFormatters();
+                .AddMvcCore();
 
             services.AddRegistration();
 
@@ -35,27 +36,35 @@ namespace GraphQlWebCore
             {
                 options.ExposeExceptions = _env.IsDevelopment();
             });
+
+            // workaround for json sync operations (used by GraphQL)
+            // iis + iisexpress
+            //services.Configure<IISServerOptions>(options =>
+            //{
+            //    options.AllowSynchronousIO = true;
+            //});
+            // kestrel
+            services.Configure<KestrelServerOptions>(options =>
+            {
+                options.AllowSynchronousIO = true;
+            });
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
 
-            app.UseMvc();
-
-
             app.UseStaticFiles();
+            app.UseRouting();
 
             // the route is just for the intro page - it is not required for the graphql service
-            app.UseMvc(routes =>
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
-                    name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                endpoints.MapControllerRoute("default", "{controller=Home}/{action=Index}/{id?}");
             });
 
             // use HTTP middleware for Schema at path /graphql
